+++ aten/src/ATen/native/native_functions.yaml
@@ -427,6 +427,10 @@
 
 - func: conv_transpose3d(Tensor input, Tensor weight, Tensor bias={}, IntList[3] stride=1, IntList[3] padding=0, IntList[3] output_padding=0, int64_t groups=1, IntList[3] dilation=1) -> Tensor
 
+- func: _copy_(Tensor self, Tensor src) -> Tensor
+  dispatch:
+    CPU: _copy__cpu
+
 - func: cos(Tensor self) -> Tensor
   variants: function, method
